requirements.txt
Flask
numpy
Pillow
matplotlib
opencv-python

